CMRNLP-103: Refactor pagination utility methods into util/pagination.py
