+++ tests/models/test_pagination.py
@@ -12,10 +12,8 @@ from models.pagination import (
     CursorParam,
     FieldsParam,
     LimitParam,
-    apply_field_filter,
-    decode_cursor,
-    encode_cursor,
 )
+from util.pagination import apply_field_filter, decode_cursor, encode_cursor
 
 
 def _b64url_encode(payload: dict) -> str:
