+++ tests/tools/get_keywords/test_get_keywords_tool.py
@@ -8,8 +8,8 @@ from unittest.mock import MagicMock, patch
 import pytest
 import requests
 
-from models.pagination import decode_cursor, encode_cursor
 from models.tools.cmr_search import SearchStatus
+from util.pagination import decode_cursor, encode_cursor
 
 
 def _load_tool() -> types.ModuleType:
