+++ tools/get_citations/tool.py
@@ -9,15 +9,13 @@ from models.pagination import (
     CursorParam,
     FieldsParam,
     LimitParam,
-    apply_field_filter,
-    encode_cursor,
-    resolve_cursor,
 )
 from models.tools.cmr_search import SearchStatus
 from models.tools.get_citations import GetCitationsInput, GetCitationsOutput
 from util.cmr.client import CMRError, search_cmr
 from util.cmr.search_tools import fetch_association_ids, normalize_citation_item
 from util.langfuse import trace_update
+from util.pagination import apply_field_filter, encode_cursor, resolve_cursor
 
 logger = logging.getLogger(__name__)
