+++ tools/get_collections/tool.py
@@ -9,9 +9,6 @@ from models.pagination import (
     CursorParam,
     FieldsParam,
     LimitParam,
-    apply_field_filter,
-    encode_cursor,
-    resolve_cursor,
 )
 from models.tools.cmr_search import SearchStatus
 from models.tools.get_collections import (
@@ -36,6 +33,7 @@ from util.cmr.search_tools import (
     normalize_collection_item,
 )
 from util.langfuse import trace_update
+from util.pagination import apply_field_filter, encode_cursor, resolve_cursor
 
 logger = logging.getLogger(__name__)
