+++ tools/get_granules/tool.py
@@ -9,9 +9,6 @@ from models.pagination import (
     CursorParam,
     FieldsParam,
     LimitParam,
-    apply_field_filter,
-    encode_cursor,
-    resolve_cursor,
 )
 from models.tools.cmr_search import SearchStatus
 from models.tools.get_granules import (
@@ -34,6 +31,7 @@ from util.cmr.search_tools import (
     normalize_granule_item,
 )
 from util.langfuse import trace_update
+from util.pagination import apply_field_filter, encode_cursor, resolve_cursor
 
 logger = logging.getLogger(__name__)
