+++ tools/get_keywords/tool.py
@@ -9,14 +9,12 @@ from models.pagination import (
     CursorParam,
     FieldsParam,
     LimitParam,
-    apply_field_filter,
-    encode_cursor,
-    resolve_cursor,
 )
 from models.tools.cmr_search import SearchStatus
 from models.tools.get_keywords import GetKeywordsInput, GetKeywordsOutput, KeywordResult
 from util.kms.client import search_kms_pattern
 from util.langfuse import trace_update
+from util.pagination import apply_field_filter, encode_cursor, resolve_cursor
 
 logger = logging.getLogger(__name__)
