+++ tools/get_services/tool.py
@@ -9,9 +9,6 @@ from models.pagination import (
     CursorParam,
     FieldsParam,
     LimitParam,
-    apply_field_filter,
-    encode_cursor,
-    resolve_cursor,
 )
 from models.tools.cmr_search import SearchStatus
 from models.tools.get_services import (
@@ -21,6 +18,7 @@ from models.tools.get_services import (
 from util.cmr.client import CMRError, search_cmr
 from util.cmr.search_tools import fetch_association_ids, normalize_service_item
 from util.langfuse import trace_update
+from util.pagination import apply_field_filter, encode_cursor, resolve_cursor
 
 logger = logging.getLogger(__name__)
